pyproject.toml
[tool.poetry]
name = "vk-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "vk_api"}]

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^7.2.0"
allure-python-commons = "^2.12.0"
jsonschema = "^4.17.3"
flake8 = "^6.0.0"
requests = "^2.28.2"
allure-pytest = "^2.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
